refactor: Move the actual applications into the root folder as the main entry points
